Move Context2D display list implementation to CustomPaintCanvas https://bugs.webkit.org/show_bug.cgi?id=261401 rdar://115277503
Reviewed by Antti Koivisto.

The display list functinality was only used for CSS Painting API in
CustomPaintCanvas. Move holding the display list into that class.
This clarifies both classes.

* Source/WebCore/WebCore.order:
* Source/WebCore/html/CustomPaintCanvas.cpp:
(WebCore::CustomPaintCanvas::getContext):
(WebCore::CustomPaintCanvas::replayDisplayList):
(WebCore::CustomPaintCanvas::copiedImage const):
(WebCore::CustomPaintCanvas::drawingContext const):
(WebCore::CustomPaintCanvas::existingDrawingContext const):
(WebCore::CustomPaintCanvas::replayDisplayListImpl):
(WebCore::CustomPaintCanvas::replayDisplayList const): Deleted.
* Source/WebCore/html/CustomPaintCanvas.h:
* Source/WebCore/html/CustomPaintImage.cpp:
(WebCore::CustomPaintImage::doCustomPaint):
* Source/WebCore/html/canvas/CanvasRenderingContext2DBase.cpp:
(WebCore::CanvasRenderingContext2DBase::reset):
(WebCore::CanvasRenderingContext2DBase::drawingContext const):
(WebCore::CanvasRenderingContext2DBase::paintRenderingResultsToCanvas): Deleted.
* Source/WebCore/html/canvas/CanvasRenderingContext2DBase.h:
(WebCore::CanvasRenderingContext2DBase::setUsesDisplayListDrawing): Deleted.

Canonical link: https://commits.webkit.org/268191@main
