+++ Source/WebCore/WebCore.order
@@ -16596,7 +16596,6 @@ __ZN7WebCore15GraphicsLayerCA19updateContentsImageEv
 __ZN7WebCore15PlatformCALayer21setMinificationFilterENS0_10FilterTypeE
 __ZN7WebCore16RenderHTMLCanvas13paintReplacedERNS_9PaintInfoERKNS_11LayoutPointE
 __ZN7WebCore17HTMLCanvasElement5paintEPNS_15GraphicsContextERKNS_10LayoutRectEb
-__ZN7WebCore22CanvasRenderingContext29paintRenderingResultsToCanvasEv
 __ZN7WebCore24setJSDOMWindowOnpopstateEPN3JSC9ExecStateEPNS0_8JSObjectENS0_7JSValueE
 __ZN7WebCore9DOMWindow13setOnpopstateEN3WTF10PassRefPtrINS_13EventListenerEEE
 __ZN7WebCore24ApplyPropertyBorderImageILNS_15BorderImageTypeE1ELNS_13CSSPropertyIDE1283EXadL_ZNKS_11RenderStyle12maskBoxImageEvEEXadL_ZNS3_15setMaskBoxImageERKNS_14NinePieceImageEEEE10applyValueES2_PNS_13StyleResolverEPNS_8CSSValueE
